# Reconfigure cells in the channel list when handling StagedChangeset.elementUpdated