### 🔗 Issue Links

Resolves [ios-issues-tracking/716](https://github.com/GetStream/ios-issues-tracking/issues/716)

### 🎯 Goal

Increase list performance and reduce UI glitches

### 📝 Summary

- Avatars flash when opening the channel list

### 🛠 Implementation

Avatars flash in the channel list when we apply `StagedChangeset.elementUpdated` changes. I observed that we call`reload(using:interrupt:setData:)` (applying data changes to the view) on the channel view 5 times after opening it (demo app) where one of the reload call contains 20 (in my case) `elementUpdated` changes. These, currently trigger cell reloads which cause the UI glitch (avatars flashing) and also reduces the performance since all the cells need to be created (it was the 4th reload call)

The proposed fix is to use Apple provided `UICollectionView.reconfigureItems(at:)` and `UITableView.reconfigureRows(at:)`. Since the channel list is a collection view and we only use a single cell type, then we can turn on cell reusing there. Message list is a table view and has more sophisticated reuse identifier handling, so currently do not use reconfiguring there (needs a separate PR since it is more complex).

I also did some micro-measurements with time profiler and iPhone 12 Pro Max.

#### Precondition
R2-D2 account is logged in.

#### Steps

1. Build the app for profiling
2. Launch the app on a device with time profiler running
3. Wait a little bit when things settle down and check what was the overall main thread usage without diving into details.

#### Results

5 consecutive runs (with 3 warmup runs before)

| Run | reloadItems (ms)  | reconfigureItems (ms) |
| --- | ------------- | ------------- |
| 1 | 484  | 381  |
| 2 | 479  | 398  |
| 3 | 499  | 387  |
| 4 | 491  | 386  |
| 5 | 494  | 393  |
| Avg | 489  | 389 |

Around 20% less CPU time on the main thread.

### 🎨 Showcase

| Before | After |
| ------ | ----- |
| ![](https://github.com/GetStream/stream-chat-swift/assets/1469907/8958a9ea-afef-465e-9d70-aaf0f536f6a5) | ![](https://github.com/GetStream/stream-chat-swift/assets/1469907/16e9c32b-79e0-472f-97f2-e1a95e752f37) |

### 🧪 Manual Testing Notes

We need to make sure that reconfiguring items does not have any side-effects. We might have relied to the fact that cells are reloaded.

### ☑️ Contributor Checklist

- [x] I have signed the [Stream CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform) (required)
- [x] This change follows zero ⚠️ policy (required)
- [ ] This change should be manually QAed
- [ ] Changelog is updated with client-facing changes
- [ ] New code is covered by unit tests
- [x] Comparison screenshots added for visual changes
- [ ] Affected documentation updated (docusaurus, tutorial, CMS)

### 🎁 Meme

![gif](https://media.giphy.com/media/v1.Y2lkPTc5MGI3NjExaDQzdDg0b3o3djV0djg3dGNueHZ0MHh0Y2hxaTJvbjF2c2tnd2Z4aSZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/1wXeZxGTdF8Uka5k4C/giphy.gif)